[[S07P22D102-65] fix : 널로 들어왔을 때 이전 이미지로 변하도록 분기, Lazy loading

See merge request s07-blockchain-contract-sub2/S07P22D102!58

+++ BE/indive/src/main/java/com/ssafy/indive/domain/member/service/MemberModifyService.java
@@ -26,21 +26,19 @@ public class MemberModifyService {
 
         findMember.update(dto);
 
-        //image만 NULL
-        if((dto.getImage() != null & !dto.getImage().isEmpty()) & (dto.getBackground()!=null & !dto.getBackground().isEmpty())){
+        if((dto.getImage() != null && !dto.getImage().isEmpty()) && (dto.getBackground()!=null && !dto.getBackground().isEmpty())){
             log.info("modifyMember : 프로필, 배경사진 둘다있음");
             findMember.uploadFiles(dto.getImage(), dto.getBackground());
-        }else if((dto.getImage() == null | dto.getImage().isEmpty()) & (dto.getBackground()==null | dto.getBackground().isEmpty())) {
+        }else if((dto.getImage() == null || dto.getImage().isEmpty()) && (dto.getBackground()==null || dto.getBackground().isEmpty())) {
             log.info("modifyMember : 프로필, 배경사진 둘다없음");
         }
-        else if((dto.getImage() == null || dto.getImage().isEmpty()) & (dto.getBackground()!=null || !dto.getBackground().isEmpty())){
+        else if((dto.getImage() == null || dto.getImage().isEmpty()) && (dto.getBackground()!=null || !dto.getBackground().isEmpty())){
             log.info("modifyMember : 프로필 없음");
             //백그라운드 넣기
             FileUtils.deleteFile(findMember.getBackgroundUuid()); // 백그라운드 초기화
             findMember.uploadBackgroundFiles(dto.getBackground());
         }
-        //background만 NULL
-        else if((dto.getImage() != null || !dto.getImage().isEmpty()) & (dto.getBackground()==null || dto.getBackground().isEmpty())) {
+        else if((dto.getImage() != null || !dto.getImage().isEmpty()) && (dto.getBackground()==null || dto.getBackground().isEmpty())) {
             log.info("modifyMember : 배경사진없음");
             //프로필 넣기
             FileUtils.deleteFile(findMember.getImageUuid());
